adaptor/requirements.txt
paho-mqtt
python-dateutil
influxdb